split add-on and fische sources with move fische to own folder
As the fische is a independent code where not access Kodi's API can it be in
own folder to have more clean and prepared for further changes.

Further becomes it easier to take the code by someone to places outside of Kodi.
