+++ src/fische/analyst.cpp
