+++ src/fische/analyst.h
