+++ src/fische/blurengine.cpp
