+++ src/fische/cpudetect.cpp
