+++ src/fische/vector.cpp
