+++ src/fische/vector.h
