+++ src/fische/vectorfield.cpp
