+++ src/fische/wavepainter.cpp
