+++ src/fische/wavepainter.h
